Uso del TAD que nos dan en metis
Creo q de estructuras nos lo dan todo hecho, habra q usar las funciones que nos dan. Tengo que hacer un main de pruebas a ver como iria la cosa

+++ src/server.c
@@ -3,7 +3,7 @@
 #define STIEMPO 30
 #define MSTIEMPO 30000
 
-int ncliente = 0;
+long ncliente = 0;
 
 int main(int argc, char *argv[]) {
